requirements.txt
gym==0.21.0
opencv-python
gym-retro
tensorflow
scikit-image
tqdm
matplotlib